Started functions for student queue!
TODO: figure out which format we're using to save the queue directly to a file and load from that file again. the student queue class is assuming that the main driver or another module is responsible for checking which place we should load the queue from (either a brand new roster or an old queue), and is responsible for providing functions to do either one.